refactor: change from edit -> update and get -> list
