+++ src/app/api/membership-type.list/route.ts
